refactor(notebook): own data execution lifecycle
